refactor: split session drawer conversation list
